3000mod/Builds/build1_oop_threaded/requirements.txt
pyserial>=3.5
PyYAML>=6.0
pytest>=7.0.0
pytest-mock>=3.10.0
